refactor (reddit_search): add types to their respective tools
