+++ mcp_servers/reddit_search/tools/__init__.py
@@ -1,10 +1,9 @@
 # Module exports and imports
-from .find_subreddits import find_relevant_subreddits
-from .search_posts import search_subreddit_posts
-from .get_comments import get_post_and_top_comments
+from .find_subreddits import find_relevant_subreddits, SubredditInfo
+from .search_posts import search_subreddit_posts, PostInfo
+from .get_comments import get_post_and_top_comments, PostDetails, CommentInfo
 from .find_similar import find_similar_posts_reddit
 from .create_post import create_post
-from .base import SubredditInfo, PostInfo, CommentInfo, PostDetails
 
 __all__ = [
     "find_relevant_subreddits",
