+++ mcp_servers/reddit_search/tools/find_subreddits.py
@@ -1,10 +1,15 @@
 import logging
-from typing import List
+from typing import List, TypedDict
 
-from .base import reddit_get, SubredditInfo
+from .base import reddit_get
 
 logger = logging.getLogger(__name__)
 
+class SubredditInfo(TypedDict):
+    """Structured data for a single subreddit."""
+    name: str
+    subscriber_count: int
+    description: str
 
 async def find_relevant_subreddits(query: str) -> List[SubredditInfo]:
     """Find subreddits that are relevant to the query and clean up the data."""
